update bubble sort
Updated the algorithm to reduce its running time. After each operation of the inner loop, the largest element is located on the right, which in the future can not be compared with the rest. Therefore, we can shift the border on the right by one - the speed of work increases by eliminating unnecessary comparisons.